Add artefact inheritance to automated detection
Automated detection will now propagate inherited artefacts into
online montage channels. The output of osl_detect_artefacts will
correspond to all of the bad segments in the active modality. That is,
inherited artefacts will be re-created as direct artefacts in the
output of this function. If the new artefacts ae discarded, the
inherited artefacts will still exist.